minimize schema.go file size
