+++ schema/query_parameter_option.go
@@ -1,0 +1,7 @@
+package schema
+
+type QueryParameterOption struct {
+	FlagName string
+	Required bool
+	Type     string
+}
